+++ app/src/main/res/layout/activity_ar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ArActivity">
+    tools:context=".ar.ArActivity">
 
     <fragment
         android:id="@+id/arimage_fragment"
