+++ app/src/main/res/layout/activity_media_input.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center"
-    tools:context=".MediaInputActivity">
+    tools:context=".ar.MediaInputActivity">
 
     <TextView
         android:id="@+id/tv_mediaHeader"
