+++ app/src/main/res/layout/activity_play.xml
@@ -6,7 +6,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PlayActivity">
+    tools:context=".sensor.PlayActivity">
 
     <Button
         android:id="@+id/button_goBack"
